# samples/microservices/order-service/src/main/java/org/berrycrush/samples/microservices/order/dto/CreateOrderRequest.java
package org.berrycrush.samples.microservices.order.dto;

import java.math.BigDecimal;

public record CreateOrderRequest(
    String customerId,
    String productId,
    Integer quantity,
    BigDecimal unitPrice
) {}