# samples/microservices/order-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: Order management microservice API
  version: 1.0.0

servers:
  - url: http://localhost:8082
    description: Local development server

paths:
  /api/orders:
    get:
      operationId: listOrders
      summary: List all orders
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    
    post:
      operationId: createOrder
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /api/orders/{id}:
    get:
      operationId: getOrder
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /api/orders/{id}/status:
    put:
      operationId: updateOrderStatus
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required:
        - customerId
        - productId
        - quantity
        - unitPrice
      properties:
        customerId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double